requirements.txt
# Core dependencies
grpcio>=1.60.0
grpcio-tools>=1.60.0
protobuf>=5.26.1,<6.0dev
uuid==1.30

# IPFS & MerkleDAG
ipfshttpclient==0.8.0a2
cid==0.1.3
multiformats==0.3.1.post4
ipld_dag_pb

# Ethereum-related dependencies
web3==6.11.3
eth-account>=0.8.0
eth-hash>=0.5.2
eth-keys>=0.4.0
eth-utils>=2.1.0
rlp>=2.0.1
pycryptodome==3.20.0

# Logging & monitoring
opentelemetry-api==1.28.0
opentelemetry-sdk==1.28.0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-instrumentation-grpc==0.49b0
prometheus-client==0.19.0
structlog==24.1.0

# Error handling & utilities
pydantic==2.6.3
rich==13.7.0

# Cobra CLI equivalent
click==8.1.7

# File utilities
watchdog==3.0.0

# YAML parsing (alternative to `go-yaml`)
PyYAML==6.0.1

#math
cryptography==36.0.0

#erasure
reedsolo
numpy

# Testing dependencies
pytest>=7.4.0
pytest-mock>=3.12.0

# Environment variables
python-dotenv
